tidied up step 1 page
